+++ backend/src/baserow/contrib/database/api/webhooks/errors.py
@@ -11,5 +11,5 @@ ERROR_TABLE_WEBHOOK_MAX_LIMIT_EXCEEDED = (
     "ERROR_TABLE_WEBHOOK_MAX_LIMIT_EXCEEDED",
     HTTP_400_BAD_REQUEST,
     f"The maximally allowed webhooks per table has been exceeded. You can create "
-    f"a maximum of {settings.WEBHOOKS_MAX_PER_TABLE} webhooks.",
+    f"a maximum of {settings.BASEROW_WEBHOOKS_MAX_PER_TABLE} webhooks.",
 )