+++ web-frontend/modules/core/utils/string.js
@@ -41,8 +41,18 @@ export const slugify = (string) => {
  * after the dot.
  */
 export const isValidURL = (str) => {
-  const pattern = /^[^\s]{0,255}(?:\.|\/\/)[^\s]{1,}$/gi
+  const pattern = /^[^\s]{0,255}(?:\.|\/\/)[^\s]{1,}$/i
   return !!pattern.test(str)
+}
+
+/**
+ * A slightly stricter URL validator than requires any url begins with a http:// or
+ * https:// and that it also passes the isValidURL validator above.
+ */
+export const isValidURLWithHttpScheme = (str) => {
+  const trimmedStr = str.trim()
+  const pattern = /^https?:\/\//i
+  return !!pattern.test(trimmedStr) && isValidURL(trimmedStr)
 }
 
 export const isValidEmail = (str) => {